Eleven: Show error message for corrupt tracks (additional codepaths)
Add showing error message for code paths where the media player fails gracefully
Unfortunately adding the logic to remove the bad track is not trivial because the
removeTracks logic automatically does other things to try to go to the next track.
For now I will leave the logic so that it goes on to the next track and make a note
on the music service rewrite notes to cover this scenario as well.

Change-Id: Ib7ae5ce7819b15cdb3ada4995bc0734ab2d9b0a2